requirements.txt
mysqlclient
Django~=3.2.5
djangorestframework~=3.12.4
markdown
django-filter
celery~=5.1.2
django-celery-results
django-celery-beat
GitPython
requests~=2.26.0
python-dotenv~=0.18.0
flake8~=3.9.2
pre-commit~=2.13.0
django-allow-cidr
